requirements_prod.txt
pyyaml>=6.0
requests>=2.31.0
google-genai>=0.3.0
notion-client>=2.0.0